requirements.txt
flask
pymongo
flask-restful
Flask-Mail
pyjwt
config
python-ldap
